test(hitl): extract shared span-capture harness and capture helper
CapturedSpans, its SpanExporter impl, and traced_as_execute_tool were
triplicated across hitl/gate.rs, hitl/tool.rs, and
mcp_tool_execution.rs; they now live once in test_span_capture, gated
on test + otel to match every consumer. mcp_tool_execution's local
captured_overrides_for is replaced by captured_overrides_multi beside
the single-pair helper it duplicated.

No test or assertion changes: 1017 aura lib tests pass unchanged.
