+++ crates/aura/src/hitl/gate.rs
@@ -289,13 +289,12 @@ mod tests {
     /// assert against.
     #[cfg(feature = "otel")]
     mod decision_id_span {
-        use std::future::Future;
-        use std::sync::atomic::{AtomicBool, Ordering};
-        use std::sync::{Arc, Mutex, MutexGuard};
 
-        use futures::future::BoxFuture;
+        use std::sync::Arc;
+        use std::sync::atomic::{AtomicBool, Ordering};
+
         use opentelemetry::trace::TracerProvider as _;
-        use opentelemetry_sdk::export::trace::{ExportResult, SpanData, SpanExporter};
+
         use opentelemetry_sdk::trace::TracerProvider;
         use rig::completion::ToolDefinition;
         use rig::tool::Tool as RigTool;
@@ -309,80 +308,8 @@ mod tests {
         use super::*;
         use crate::approval_event_broker::{self, ApprovalLifecycleEvent};
         use crate::logging::ATTR_DECISION_ID;
+        use crate::test_span_capture::{CapturedSpans, traced_as_execute_tool};
         use crate::tool_wrapper::WrappedTool;
-
-        /// Spans the test subscriber has exported.
-        #[derive(Debug, Clone, Default)]
-        struct CapturedSpans(Arc<Mutex<Vec<SpanData>>>);
-
-        impl CapturedSpans {
-            fn spans(&self) -> MutexGuard<'_, Vec<SpanData>> {
-                self.0.lock().expect("captured spans mutex")
-            }
-
-            fn contains(&self, name: &str) -> bool {
-                self.spans().iter().any(|span| span.name == name)
-            }
-
-            /// The single `key` attribute on the span named `name`, or `None`
-            /// if the span carries no such attribute. Panics if the span
-            /// carries more than one — a double-stamp regression would
-            /// otherwise pass with only the first entry.
-            fn attribute(&self, name: &str, key: &str) -> Option<String> {
-                let spans = self.spans();
-                let span = spans.iter().find(|span| span.name == name)?;
-                let matches: Vec<_> = span
-                    .attributes
-                    .iter()
-                    .filter(|kv| kv.key.as_str() == key)
-                    .collect();
-                assert!(
-                    matches.len() <= 1,
-                    "span {name:?} must carry at most one {key} attribute, found {}: \
-                     a regression is double-stamping the same span",
-                    matches.len(),
-                );
-                matches.first().map(|kv| kv.value.to_string())
-            }
-        }
-
-        impl SpanExporter for CapturedSpans {
-            fn export(&mut self, batch: Vec<SpanData>) -> BoxFuture<'static, ExportResult> {
-                self.spans().extend(batch);
-                Box::pin(std::future::ready(Ok(())))
-            }
-        }
-
-        /// Run `body` inside an `execute_tool` span — the span Rig opens around
-        /// a tool call — under a subscriber that exports to memory, returning
-        /// the body's output and the `decision_id` the exported span carries.
-        async fn traced_as_execute_tool<T>(body: impl Future<Output = T>) -> (T, Option<String>) {
-            let captured = CapturedSpans::default();
-            let provider = TracerProvider::builder()
-                .with_simple_exporter(captured.clone())
-                .build();
-            let _guard = tracing::subscriber::set_default(
-                tracing_subscriber::registry()
-                    .with(tracing_opentelemetry::layer().with_tracer(provider.tracer("test"))),
-            );
-
-            let output = body.instrument(tracing::info_span!("execute_tool")).await;
-
-            // The registry instruments a parked approval's wake task with the
-            // same span, so the export lands once that task has released it too.
-            for _ in 0..1_000 {
-                if captured.contains("execute_tool") {
-                    break;
-                }
-                tokio::task::yield_now().await;
-            }
-            assert!(
-                captured.contains("execute_tool"),
-                "the execute_tool span was never exported",
-            );
-
-            (output, captured.attribute("execute_tool", ATTR_DECISION_ID))
-        }
 
         /// What a trace backend actually receives, assembled the way the binary
         /// assembles it: the real OTel filter, the OpenInference exporter, and
